bpo-38632: Allow to pass timestamp values when creating a tgz
Gzip embeds current time in its header which is inconveninent when
trying to do bytes for bytes reproducible builds with SOURCE_DATE_EPOCH.

This allow to pass timestamp to various constructor in the tarfile
module which will be used – when present – by the gzip module to set the
correct time, among many things in the header.

As far as I can tell gzip is the only compression method that gives
time-dependant results, and/or have exposed options to set this time.